+++ doc/version_history.rst
@@ -14,6 +14,7 @@ Changes:
 * Changed the ApSCS positionActual from one to two values.
 * Added type annotations.
 * Rewrote the way the JSON schemas are loaded.
+* Updated the error part of the AMCS status reply.
 
 Requires: